+++ packages/@justweb3/ui/src/lib/components/ClickableItem/index.tsx
@@ -8,7 +8,7 @@ interface ClickableListItemProps {
   subtitle?: string | React.ReactNode;
   left?: React.ReactNode;
   loading?: boolean;
-  onClick?: () => void;
+  onClick?: (e: React.MouseEvent<HTMLDivElement>) => void;
   disabled?: boolean;
   right?: React.ReactNode;
   onHover?: (hover: boolean) => void;
@@ -52,10 +52,12 @@ export const ClickableItem = React.forwardRef<
       <div
         ref={ref}
         className={clsx(wrapperClassNames, className)}
-        onClick={() => {
+        onClick={(e) => {
           if (!loading && !disabled && clickable) {
-            onClick && onClick();
+            onClick && onClick(e);
           }
+          // e.stopPropagation();
+          // e.preventDefault();
         }}
         onPointerEnter={() => {
           if (!loading && !disabled && clickable) {