+++ packages/@justweb3/ui/src/lib/ui/Checkbox/index.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import * as CheckboxPrimitive from '@radix-ui/react-checkbox';
-import { Check } from 'lucide-react';
 import styles from './Checkbox.module.css';
 
 const Checkbox = React.forwardRef<
@@ -13,7 +12,18 @@ const Checkbox = React.forwardRef<
     {...props}
   >
     <CheckboxPrimitive.Indicator className={styles.indicator}>
-      <Check className={styles.icon} />
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 24 24"
+        fill="none"
+        stroke="currentColor"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        className={styles.icon}
+      >
+        <path d="M20 6 9 17l-5-5"></path>
+      </svg>
     </CheckboxPrimitive.Indicator>
   </CheckboxPrimitive.Root>
 ));