+++ packages/@justweb3/xmtp-plugin/src/lib/components/ChatSheet/index.tsx
@@ -1,4 +1,6 @@
 import {
+  AddIcon,
+  Flex,
   Sheet,
   SheetContent,
   SheetTitle,
@@ -7,7 +9,6 @@ import {
   TabsList,
   TabsTrigger,
 } from '@justweb3/ui';
-import React, { useEffect, useMemo } from 'react';
 import {
   CachedConversation,
   ContentTypeMetadata,
@@ -16,20 +17,23 @@ import {
   useStreamAllMessages,
   useStreamConversations,
 } from '@xmtp/react-sdk';
+import React, { useEffect, useMemo } from 'react';
 import { ChatList } from '../ChatList';
 
 export interface ChatSheetProps {
   open?: boolean;
   handleOpen?: (open: boolean) => void;
   handleOpenChat: (
-    conversation: CachedConversation<ContentTypeMetadata> | null
+    conversation: CachedConversation<ContentTypeMetadata>
   ) => void;
+  handleNewChat: () => void;
 }
 
 export const ChatSheet: React.FC<ChatSheetProps> = ({
   open,
   handleOpen,
   handleOpenChat,
+  handleNewChat,
 }) => {
   const [tab, setTab] = React.useState('Chats');
   const { conversations, isLoading } = useConversations();
@@ -74,6 +78,28 @@ export const ChatSheet: React.FC<ChatSheetProps> = ({
     <Sheet open={open} onOpenChange={handleOpen}>
       <SheetContent side="right" overlay={false} style={{ width: '100%' }}>
         <SheetTitle>Chats</SheetTitle>
+        <Flex
+          align="center"
+          justify="center"
+          style={{
+            width: 45,
+            height: 45,
+            borderRadius: '50%',
+            backgroundColor: 'var(--justweb3-primary-color',
+            cursor: 'pointer',
+            position: 'absolute',
+            bottom: '2rem',
+            right: '2rem',
+            zIndex: 100,
+          }}
+        >
+          <AddIcon
+            onClick={handleNewChat}
+            fill={'var(--justweb3-background-color'}
+            width={35}
+            height={35}
+          />
+        </Flex>
         <Tabs
           defaultValue={'Chats'}
           value={tab}
@@ -83,6 +109,7 @@ export const ChatSheet: React.FC<ChatSheetProps> = ({
             flexDirection: 'column',
             marginBottom: '0px',
             overflow: 'hidden',
+            marginTop: '10px',
             flex: '1',
           }}
         >
@@ -95,9 +122,29 @@ export const ChatSheet: React.FC<ChatSheetProps> = ({
             </TabsTrigger>
             <TabsTrigger
               value={'Requests'}
-              style={{ flexBasis: 'calc( 100% / 3)' }}
+              style={{ flexBasis: 'calc( 100% / 3)', position: 'relative' }}
             >
               Requests
+              {requestConversations.length > 0 && (
+                <Flex
+                  justify="center"
+                  align="center"
+                  style={{
+                    position: 'absolute',
+                    top: 0,
+                    right: 10,
+                    color: 'var(--justweb3-background-color)',
+                    background: 'var(--justweb3-primary-color)',
+                    width: 17,
+                    height: 17,
+                    borderRadius: '50%',
+                    lineHeight: 0.5,
+                    fontSize: '10px',
+                  }}
+                >
+                  {requestConversations.length}
+                </Flex>
+              )}
             </TabsTrigger>
             <TabsTrigger
               value={'Blocked'}
@@ -110,22 +157,23 @@ export const ChatSheet: React.FC<ChatSheetProps> = ({
             <div>Loading...</div>
           ) : (
             <>
-              <TabsContent value={'Chats'}>
+              <TabsContent value={'Chats'} style={{ overflowY: 'scroll' }}>
                 <ChatList
                   conversations={allowedConversations}
                   handleOpenChat={handleOpenChat}
                 />
               </TabsContent>
-              <TabsContent value={'Requests'}>
+              <TabsContent value={'Requests'} style={{ overflowY: 'scroll' }}>
                 <ChatList
                   conversations={requestConversations}
                   handleOpenChat={handleOpenChat}
                 />
               </TabsContent>
-              <TabsContent value={'Blocked'}>
+              <TabsContent value={'Blocked'} style={{ overflowY: 'scroll' }}>
                 <ChatList
                   conversations={blockedConversations}
                   handleOpenChat={handleOpenChat}
+                  blockedList
                 />
               </TabsContent>
             </>